rename client to cli
